Mercury/requirements.txt
aniso8601==7.0.0
apiclient==1.0.4
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
coverage==5.0.4
crypto==1.4.1
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==2.2.8
django-appconf==1.0.3
django-ckeditor==5.7.1
django-cors-headers==3.1.1
django-crispy-forms==1.9.0
django-easy-select2==1.5.6
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-imagekit==4.0.2
django-import-export==1.2.0
django-js-asset==1.2.2
djangorestframework==3.10.3
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
graphene==2.1.8
graphene-django==2.6.0
graphene-django-extras==0.4.8
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
jdcal==1.4.1
MarkupPy==1.14
Naked==0.1.31
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.0
pbr==5.4.4
pilkit==2.0
Pillow==6.2.1
promise==2.2.1
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==1.6.1
service-identity==18.1.0
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
stevedore==1.32.0
tablib==0.14.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==20.0.13
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
xlrd==1.2.0
xlwt==1.3.0
youtube-transcript-api==0.3.1
zipp==3.1.0
zope.interface==5.0.2
